Moved SelectorManager class to pick.py
